:lipstick: Update: convert styled-component to React component #125
