+++ taskiano/frontend/components/ClockTimer/styles.module.scss
@@ -1,0 +1,20 @@
+@import "../../styles/variables.scss";
+
+.clockTimerContainer {
+  @include flex-box($justify: center, $align: center);
+
+  width: 1.5rem;
+  height: 1.5rem;
+  border-radius: 50%;
+}
+
+.calendarTimeContainer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  width: 1.5rem;
+  height: 1.5rem;
+
+  color: $gray;
+}
